some cosmetic changes in percentages representaions for reports service

+++ src/TisaBackend/TisaBackend.Domain/Models/FlightReportData.cs
@@ -4,6 +4,6 @@
     {
         public int FlightId { get; set; }
         public int TotalOfIncome { get; set; }
-        public double OccupancyPercentage { get; set; }
+        public int OccupancyPercentage { get; set; }
     }
 }
